Testing the code in a Jupyter Notebook to run with python
Jonathan discovered we need a separate file after dealing with an error message "attempted relative import with no known parent package" typically occurs when you're trying to run a Python file as a script, which is part of a package, from outside the package. In your case, the parent package is quantstats. The .. in the import statement is a relative import that refers to the parent package.